+++ lib/black_thursday_helper.rb
@@ -25,18 +25,6 @@ module BlackThursdayHelper
    end
   end
 
-  def update(id, attributes)
-      if find_by_id(id) != nil
-      object_to_be_updated = find_by_id(id)
-      object_to_be_updated.name = attributes[:name]
-      object_to_be_updated.description = attributes[:description]
-      object_to_be_updated.unit_price = attributes[:unit_price]
-      object_to_be_updated.updated_at = Time.now
-      else
-        nil
-      end
-  end
-
   def delete(id)
     @collections.delete_if do |object|
       object.id == id
