+++ lib/items_repo.rb
@@ -50,4 +50,16 @@ class ItemsRepo
     item
   end
 
+  def update(id, attributes)
+      if find_by_id(id) != nil
+      object_to_be_updated = find_by_id(id)
+      object_to_be_updated.name = attributes[:name]
+      object_to_be_updated.description = attributes[:description]
+      object_to_be_updated.unit_price = attributes[:unit_price]
+      object_to_be_updated.updated_at = Time.now
+      else
+        nil
+      end
+  end
+
 end
